% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm.lik.R
\name{bm.lik}
\alias{bm.lik}
\title{BM likelihood function}
\usage{
bm.lik(cache, p, method)
}
\arguments{
\item{cache}{A cache object created using the cache.epc function}

\item{p}{Numeric vector of starting parameters to be loaded corresponding to theta and sig2. If NULL start.searcher function will be run instead.}

\item{method}{A string corresponding to the optimx algorithm to be run, recommended 'Nelder-Mead' for first search followed by 'BFGS' for the second.}
}
\value{
A table of parameters (theta and sig2) with a likelihood, execution time, model convergence parameters from optimx::optimx (ignore), and AIC.
}
\description{
Performs a single maximum likelihood search under an Brownian Motion process
}
