% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.epc.R
\name{cache.epc}
\alias{cache.epc}
\title{EPC cache building function}
\usage{
cache.epc(
  mtree,
  epc_params,
  m_type,
  environment_data,
  trait_data,
  n_slice,
  env_treat = "bin",
  custom.model.t = NULL,
  custom_params_t = NULL,
  custom.model.a = NULL,
  custom_params_a = NULL,
  custom_param_names = NULL
)
}
\arguments{
\item{mtree}{A phylogenetic tree}

\item{epc_params}{EPC params included in model. Values can be 'alpha' , 'theta', 'OU', or 'BM'. 'BM' and 'OU' are to simulate Brownian motion or a single peak Ornstein-Uhlenbeck model, respectively.}

\item{m_type}{A string corresponding to the EPC environment relationship, can be 'linear', 'step', or 'custom'.}

\item{environment_data}{A two column table where column 1 is time and column 2 is the environment. Time is ordered such that 0 corresponds to the present and larger values indicate further back in time.}

\item{trait_data}{A vector of numeric trait data where each value has a corresponding species name}

\item{n_slice}{Number of time slices across the tree, must equal the 'X' parameter.}

\item{custom.model.t}{A user specified function that outputs a value of theta given some environmental value.}

\item{custom_params_t}{Parameters to simulate for a custom EPC theta relationship given 'custom.model.t'.}

\item{custom.model.a}{A user specified function that outputs a value of alpha given some environmental value.}

\item{custom_params_a}{Parameters to simulate for a custom EPC alpha relationship given 'custom.model.a'.}

\item{custom_param_names}{Names of custom parameters for custom EPC - alpha/theta models.}
}
\value{
A cache object that includes the information needed for a maximum likelihood search.
}
\description{
Simulates traits under an EPC process for one or more OU model parameters
}
