% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_aic_weights.R
\name{calculate_aic_weights}
\alias{calculate_aic_weights}
\title{AIC weight calculator}
\usage{
calculate_aic_weights(aic_df)
}
\arguments{
\item{aic_df}{A dataframe of AIC scores where each column is a model type and each row a simulation. AIC weights will be presented row wise.}
}
\value{
An AIC weight table.
}
\description{
AIC weight calculator
}
