% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consist.sim.R
\name{consist.sim}
\alias{consist.sim}
\title{A parametric bootstrap function for results of an EPC maxmimum likelihood search}
\usage{
consist.sim(params, cache, reps, cores, c_t = NULL, c_a = NULL)
}
\arguments{
\item{params}{A single numeric row of parameter results from epc.max.lik,}

\item{cache}{An EPC cache object used for epc.max.lik}

\item{reps}{Number of simulations to perform}

\item{cores}{Number of cores to use for simulations, do not exceed your machines limits!}
}
\value{
A list with maximum likelihood search results for each reptiion under all model types and all results weighted by AIC weights.
}
\description{
Given parameter results from an epc.max.lik search and an cache.epx object, perform a parametric bootstrap. Returns a AIC weight table corresponding to each model.
}
