% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addshift2map.R
\name{.addshift2map}
\alias{.addshift2map}
\title{Support function to load pars for epoch time slicing taken from bayou}
\usage{
.addshift2map(x, maps = maps, sb = sb, loc = loc, t2 = t2)
}
\arguments{
\item{x}{tree}
}
\description{
Support function to load pars for epoch time slicing taken from bayou
}
