% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epc_utilities.R
\name{.D.from.theta}
\alias{.D.from.theta}
\title{bayOU internal function.}
\usage{
.D.from.theta(pars, cache, sort = TRUE)
}
\description{
\code{.heights.cache} is an internal function and not generally called by the user
}
\details{
This function calculates the change in optima at each shift point on the tree
}
