% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeHeights.R
\name{.nodeHeights}
\alias{.nodeHeights}
\title{Support function to get node heights}
\usage{
.nodeHeights(tree, scale = "YBP")
}
\arguments{
\item{x}{tree}
}
\description{
Support function to get node heights
}
