% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epc_utilities.R
\name{.repars}
\alias{.repars}
\title{Support functions from bayou and other sources}
\usage{
.repars(pars, expected)
}
\arguments{
\item{x}{tree}
}
\description{
Support functions from bayou and other sources
}
