% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.epoch.fossil.R
\name{drop.epoch.fossil}
\alias{drop.epoch.fossil}
\title{Drop all tips from a time epoch}
\usage{
drop.epoch.fossil(phy, m, M, tol = 1e-08)
}
\arguments{
\item{phy}{A phylogenetic tree.}

\item{m}{Lower bound of time to drop tips from}

\item{M}{Upper bound of time to drop tips from}
}
\value{
A tree with dropped extant tips
}
\description{
Drop all tips from a time epoch
}
