% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.random.fossil.R
\name{drop.random.fossil}
\alias{drop.random.fossil}
\title{Drop a set of random fossil tips}
\usage{
drop.random.fossil(phy, m, tol = 1e-08)
}
\arguments{
\item{phy}{A phylogenetic tree.}

\item{m}{Number of tips to drop}
}
\value{
A tree with dropped extant tips
}
\description{
Drop a set of random fossil tips
}
