% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epc.dent.R
\name{epc.dent}
\alias{epc.dent}
\title{EPC dentist search function}
\usage{
epc.dent(cache, pars, best_log, c_t = NULL, c_a = NULL, reps = NULL)
}
\arguments{
\item{cache}{A cache object created using epcTools.}

\item{pars}{Numeric vector of starting parameters to be loaded corresponding to parameters specified in cache object (format: theta, sig2, alpha; if any parameter varies with the environment the model parameters follow the same order). If NULL start.searcher function will be run instead.}

\item{best_log}{Best likelihood value for dentist to search around}

\item{c_t}{A custom theta function under an EPC process, NULL if linear or step is being used.}

\item{c_a}{A custom alpha function under an EPC process, NULL if linear or step is being used.}
}
\value{
Dentist output with 95% confidence interval values for all parameters.

Note: This function can run into recursion limits on individual systems which may ultimately prevent it's use by epc.max.lik. Use the skip.dentist=TRUE flag on epc.max.lik if you run into this issue.
}
\description{
Performs dentist search for all model parameters for BM, OU, and EPC models
}
