% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epc.max.lik.R
\name{epc.max.lik}
\alias{epc.max.lik}
\title{EPC maximum likelihood multisearch function}
\usage{
epc.max.lik(cache, p, c_t = NULL, c_a = NULL, skip_dentist = FALSE)
}
\arguments{
\item{cache}{A cache object created using epcTools.}

\item{p}{Numeric vector of starting parameters to be loaded corresponding to parameters specified in cache object (format: theta, sig2, alpha; if any parameter varies with the environment the model parameters follow the same order). If NULL start.searcher function will be run instead.}

\item{c_t}{A custom theta function under an EPC process, NULL if linear or step is being used.}

\item{c_a}{A custom alpha function under an EPC process, NULL if linear or step is being used.}
}
\value{
A list including a table of parameters (theta,sig2, and alpha or any combination of EPC model parameters) with likelihood, execution time, model convergence parameters from optimx::optimx (ignore), and AIC. Dentist output with 95% confidence interval values for all parameters.
}
\description{
Performs repeated maximum likelihood searches under an EPC process for one or more OU model parameters. Then, performs a dentist run to determine 95% CI for model parameters.
}
