% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epc.var.plot.R
\name{epc.var.plot}
\alias{epc.var.plot}
\title{EPC environment-trait relationship visualization function}
\usage{
epc.var.plot(max_output, cache)
}
\arguments{
\item{max_output}{Output from epc.max.lik}

\item{cache}{An EPC cache object.}
}
\value{
An epc.var.plot that depicts the optima and stationary variance shifting through time under the maximum likelihood estimate
}
\description{
Visualizes epc.max.lik outputs and reconstructed relationships such as the stationary variance through time or shifts in optimal values
}
