% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epcTools.R
\docType{package}
\name{epcTools}
\alias{epcTools}
\title{epcTools Package}
\description{
Given a tree, an environmental curve, and trait values, this package permits OU model parameters to vary with the environmental curve under a defined relationship (e.g. linear, step, custom).
}
\author{
T. Mason Linscott \email{mason.linscott@gmail.com}
}
