% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBranchesSlice.R
\name{getBranchesSlice}
\alias{getBranchesSlice}
\title{Basic likelihood function}
\usage{
getBranchesSlice(slice, tree, nH1, nH2)
}
\arguments{
\item{slice}{A vector of slice times for the tree.}

\item{trees}{A phylogenetic tree.}
}
\value{
A sliced tree for loading into epoch simmap
}
\description{
Returns a likelihood value given a cache object and vector of parameters
}
