% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSliceEnv.R
\name{getSliceEnv}
\alias{getSliceEnv}
\title{Environmental spline function}
\usage{
getSliceEnv(env, b)
}
\arguments{
\item{env}{A dataframe of environment data and time}

\item{b}{Number of time slices}

\item{R}{Resolution of spline}
}
\value{
A numeric likelihood value
}
\description{
Given a dataframe of environment values at a given time, return a vector of length b that corresponds to environmental values at a time slice. This can be done through a loess smoothed function or from taking the average of values at a given time interval (binning).
}
