% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik.pars2pcmbase.R
\name{lik.pars2pcmbase}
\alias{lik.pars2pcmbase}
\title{Function for loading parameters into pcmbase format}
\usage{
lik.pars2pcmbase(pars, cache)
}
\arguments{
\item{pars}{Numeric vector of parameters to be loaded.}

\item{cache}{A cache object created using epcTools.}
}
\value{
List of pars for loading into pcmbase.
}
\description{
Function for loading parameters into pcmbase format
}
