% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEpochSimmap.R
\name{makeEpochSimmap}
\alias{makeEpochSimmap}
\title{Make epoch simmaps}
\usage{
makeEpochSimmap(tree, slices)
}
\arguments{
\item{tree}{A phylogenetic tree.}

\item{slices}{A vector where tree will be cut into slices}
}
\value{
An epoch simmap tree
}
\description{
Cuts a tree up into epochs with the same mapped simmap state across the tree
}
\details{
From Josef Uyeda
}
