% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars3simmap.R
\name{pars3simmap}
\alias{pars3simmap}
\title{Support function to load pars for epoch time slicing}
\usage{
pars3simmap(pars, tree)
}
\arguments{
\item{pars}{Numeric vector of parameters to be used to determine simmap positions}

\item{tree}{A phylogenetic tree.}
}
\value{
Vector of values to be passed for simmap creation
}
\description{
Support function to load pars for epoch time slicing
}
