% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dentist.R
\name{plot.epc}
\alias{plot.epc}
\title{Plot dentist parameter outputs}
\usage{
\method{plot}{epc}(epc_out, local.only = FALSE, ...)
}
\arguments{
\item{epc_out}{An epc.max.lik output to be plotted}

\item{local.only}{If only the 95% CI should be plotted.}
}
\value{
A plot corresponding to plot.dentist with 95% confidence values for each parameter plotted.
}
\description{
A lightly modified plot.dentist function for the epcTools framework
}
