% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.epc.phenogram.R
\name{sim.epc.phenogram}
\alias{sim.epc.phenogram}
\title{Plot simulated EPC phenogram}
\usage{
sim.epc.phenogram(cache, n)
}
\arguments{
\item{n}{Number of times to simulate and plot the data.}

\item{caches}{A cache object created using epcTools.}
}
\value{
A phenogram plot of trait values through time.
}
\description{
A plot function that will plot a phenogram for simulated EPC cache objects
}
