% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.EPC.R
\name{sim.EPC}
\alias{sim.EPC}
\title{EPC simulation function}
\usage{
sim.EPC(
  mtree,
  epc_params,
  m_type,
  X,
  n_slice,
  reps_per_tree = 1,
  theta = NULL,
  sig2,
  alpha = NULL,
  b0_a = NULL,
  b1_a = NULL,
  b0_t = NULL,
  b1_t = NULL,
  step_thresh_t = NULL,
  step_high_t = NULL,
  step_low_t = NULL,
  step_thresh_a = NULL,
  step_high_a = NULL,
  step_low_a = NULL,
  custom.model.t = NULL,
  custom_params_t = NULL,
  custom.model.a = NULL,
  custom_params_a = NULL,
  custom_param_names = NULL
)
}
\arguments{
\item{mtree}{A phylogenetic tree}

\item{epc_params}{EPC params included in model. Values can be 'alpha' , 'theta', 'OU', or 'BM'. 'BM' and 'OU' are to simulate non EPC models Brownian motion and Ornstein-Uhlenbeck, respectively.}

\item{m_type}{A string corresponding to the EPC environment relationship, can be 'linear', 'step', or 'custom'.}

\item{X}{A vector corresponding to environmental values at each time slice}

\item{n_slice}{Number of time slices across the tree, must equal the 'X' parameter.}

\item{reps_per_tree}{Number of times to simulate data under the given parameters.}

\item{theta}{A numeric value for theta for simulation.}

\item{sig2}{A numeric value for sig2 for simulation.}

\item{alpha}{A numeric value for alpha for simulation.}

\item{b0_a}{A numeric value for the intercept b0 under a linear relationship between the environment and alpha.}

\item{b1_a}{A numeric value for the slope b1 under a linear relationship between the environment and alpha.}

\item{b0_t}{A numeric value for the intercept b0 under a linear relationship between the environment and theta.}

\item{b1_t}{A numeric value for the slope b1 under a linear relationship between the environment and theta.}

\item{step_thresh_t}{A numeric value for the threshold of a step relationship between the environment and theta.}

\item{step_high_t}{A numeric value of theta when the environment exceeds step_thresh_t'.}

\item{step_low_t}{A numeric value of theta when the environment falls below step_thresh_t'.}

\item{step_thresh_a}{A numeric value for the threshold of a step relationship between the environment and alpha.}

\item{step_high_a}{A numeric value of alpha when the environment exceeds step_thresh_a'.}

\item{step_low_a}{A numeric value of alpha when the environment falls below step_thresh_a'.}

\item{custom.model.t}{A user specified function that outputs a value of theta given some environmental value.}

\item{custom_params_t}{Parameters to simulate for a custom EPC theta relationship given 'custom.model.t'.}

\item{custom.model.a}{A user specified function that outputs a value of alpha given some environmental value.}

\item{custom_params_a}{Parameters to simulate for a custom EPC alpha relationship given 'custom.model.at'.}

\item{custom_param_names}{Names of custom parameters for custom EPC - alpha/theta models.}
}
\value{
A cache object that includes the information needed for a maximum likelihood search.
}
\description{
Simulates traits under an EPC process for one or more OU model parameters
}
