% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTree.R
\name{simTree}
\alias{simTree}
\title{Simulate a phylogeny}
\usage{
simTree(n_extant, tree_size, max_time, n_tree)
}
\arguments{
\item{n_extant}{Number of extant tips.}

\item{tree_size}{Size of tree to simulate}

\item{max_time}{Max length of tree desired in time}

\item{n_tree}{Number of trees to simulate}
}
\value{
A phylogeny
}
\description{
Simulate a phylogeny
}
