% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.searcher.R
\name{start.searcher}
\alias{start.searcher}
\title{Starting parameter searcher}
\usage{
\method{start}{searcher}(cache)
}
\arguments{
\item{cache}{A cache object created using epcTools.}
}
\value{
A vector of starting parameters corresponding to information in the cache object.
}
\description{
Finds starting parameters for BM, OU, and EPC models
}
