% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.epc.R
\name{summary.epc}
\alias{summary.epc}
\title{Summarize EPC run}
\usage{
\method{summary}{epc}(epc_out)
}
\arguments{
\item{epc_out}{An epc.max.lik output to be plotted}
}
\value{
A summary of parameters and model performance for an epc.max.lik object
}
\description{
A summary function for the output of epc.max.lik
}
