% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.tree.maker.R
\name{ts.tree.maker}
\alias{ts.tree.maker}
\title{EPC environment-trait relationship visualization function}
\usage{
ts.tree.maker(TS)
}
\arguments{
\item{TS}{A time series}
}
\value{
A phylogeny constructed from the time series.
}
\description{
Visualizes epc.max.lik outputs and reconstructed relationships such as the stationary variance through time or shifts in optimal values
}
