% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consist.sim.R
\name{consist.sim}
\alias{consist.sim}
\title{Consistency simulator of EPC results}
\usage{
consist.sim(params, cache, reps, cores, c_t = NULL, c_a = NULL)
}
\arguments{
\item{params}{A single numeric row of parameter results from epc.max.lik,}

\item{cache}{An EPC cache object used for epc.max.lik}

\item{reps}{Number of simulations to perform}

\item{cores}{Number of cores to use for simulations, do not exceed your computers limits!}
}
\value{
A list with maximum likelihood search results for each reptition under all model types and all results weighted by AIC weights.
}
\description{
Given parameter results from an epc.max.lik, simulate many EPC cache objects with same parameters to determine EPC model identifiability
}
