% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.random.extant.R
\name{drop.random.extant}
\alias{drop.random.extant}
\title{Drop a random set of extant tips}
\usage{
drop.random.extant(phy, m, tol = 1e-08)
}
\arguments{
\item{phy}{A phylogenetic tree.}

\item{m}{Number of tips to drop}
}
\value{
A tree with dropped extant tips
}
\description{
Drop a random set of extant tips
}
